Fixed all the other tests for the Builders.
